+++ contentcuration/contentcuration/urls.py
@@ -232,7 +232,7 @@ urlpatterns += [
     url(r'^zipcontent/(?P<zipped_filename>[^/]+)/(?P<embedded_filepath>.*)', zip_views.ZipContentView.as_view(), {}, "zipcontent"),
     url(r'^api/file_upload/', file_views.file_upload, name="file_upload"),
     url(r'^api/file_create/', file_views.file_create, name="file_create"),
-    url(r'^api/generate_thumbnail/$', file_views.generate_thumbnail, name='generate_thumbnail'),
+    url(r'^api/generate_thumbnail/(?P<contentnode_id>[^/]*)$', file_views.generate_thumbnail, name='generate_thumbnail'),
 ]
 
 # Add account/registration endpoints
